requirements.txt
djangorestframework~=3.14.0
python-dotenv~=1.0.1
MergePythonClient
Django~=5.0.1
urllib3==2.0.7
requests==2.31.0
pylint==3.0.3